reuse http client, move module
