+++ api/app/api/dependencies.py
@@ -1,4 +1,3 @@
-import httpx
 from fastapi import Request
 
 from app.core.database import DatabaseSessionManager
@@ -10,11 +9,10 @@ from app.services.place.repository import PlaceRepository
 
 async def get_geocoding_service(request: Request) -> GeocodingService:
     settings: Settings = request.app.state.settings
-    async with httpx.AsyncClient() as http_client:
-        yield GeocodingService(
-            api_key=settings.geoapify_api_key,
-            http_client=http_client,
-        )
+    yield GeocodingService(
+        api_key=settings.geoapify_api_key,
+        http_client=request.app.state.http_client,
+    )
 
 
 async def get_coffee_repo(request: Request) -> PlaceRepository:
